Moved conversion to snake case function from apc to common Mapping now converts dwc field to snake case for better matching to back end Updated ncbi refseq/genbank augment to use snake case
