+++ src/lib/commonFuncs.py
@@ -105,3 +105,18 @@ def clearFolder(folderPath: Path, delete: bool = False) -> None:
     
     if delete:
         folderPath.rmdir()
+
+def toSnakeCase(s: str) -> str:
+    ret = ""
+    upperCount = 0
+    for char in s:
+        if char.isalpha() and char.upper() == char:
+            if upperCount == 0:
+                ret += "_"
+            ret += char.lower()
+            upperCount += 1
+        else:
+            upperCount = 0
+            ret += char
+
+    return ret
